Backend Changes for LMS

+++ Services/EmployeeService.cs
@@ -1,6 +1,9 @@
-﻿using LeaveManagementSystem_Backend.IRepository;
+﻿using LeaveManagementSystem_Backend.DBContext;
+using LeaveManagementSystem_Backend.IRepository;
 using LeaveManagementSystem_Backend.IServices;
 using LeaveManagementSystem_Backend.Models;
+using LeaveManagementSystem_Backend.Utilities;
+using Microsoft.EntityFrameworkCore;
 
 namespace LeaveManagementSystem_Backend.Services
 {
@@ -13,25 +16,76 @@ namespace LeaveManagementSystem_Backend.Services
         {
             _employeeRepository = employeeRepository;
             _allocatedLeaveService = allocatedLeaveService;
+           
         }
-        public async Task<Employee> CreateEmployee(Employee employee)
+        public async Task<Employee> CreateEmployee(EmployeeRequest employeeRequest)
         {
-            var newEmployee = await _employeeRepository.CreateEmployee(employee);
+            var newEmployee = new Employee
+            {
+                FirstName = employeeRequest.FirstName,
+                LastName = employeeRequest.LastName,
+                Email = employeeRequest.Email,
+                JoinDate = employeeRequest.JoinDate,
+                EmployeeNumber = employeeRequest.EmployeeNumber,
+                NicNo = employeeRequest.NicNo,
+                Dob = employeeRequest.Dob,
+                Nationality = employeeRequest.Nationality,
+                MaritalStatus = employeeRequest.MaritalStatus,
+                PositionId = employeeRequest.PositionId,
+                TelephoneNumber = employeeRequest.TelephoneNumber,
+                MobileNumber = employeeRequest.MobileNumber,
+                PermanentAddress = employeeRequest.PermanentAddress,
+                CurrentAddress = employeeRequest.CurrentAddress,
+                EmergencyContactName = employeeRequest.EmergencyContactName,
+                EmergencyContactNumber = employeeRequest.EmergencyContactNumber,
+                EmergencyContactRelationship = employeeRequest.EmergencyContactRelationship,
+                BankName = employeeRequest.BankName,
+                AccountNo = employeeRequest.AccountNo,
+                AccountHolder = employeeRequest.AccountHolder,
+                Branch = employeeRequest.Branch,
+                TypeOfAccount = employeeRequest.TypeOfAccount,
+                Profile = employeeRequest.Profile,
+                ConfirmDate = employeeRequest.ConfirmDate,
+                //Status = employeeRequest.Status,
+                
+            };
+
+            var createdEmployee = await _employeeRepository.CreateEmployee(newEmployee);
 
             try
             {
-                await _allocatedLeaveService.AllocateAnnualLeave(newEmployee);
-                await _allocatedLeaveService.AllocatedCasualLeave(newEmployee);
+                await _allocatedLeaveService.AllocateAnnualLeave(createdEmployee);
+                await _allocatedLeaveService.AllocatedCasualLeave(createdEmployee);
+
+                if (employeeRequest.CreateUser)
+                {
+                 
+                    var autoPassword = PasswordUtility.GenerateAutoPassword();
+                    var user = new User
+                    {
+                        EmployeeId = createdEmployee.Id,
+                        Email = createdEmployee.Email,
+                        PasswordHash = PasswordUtility.HashPassword(autoPassword)
+                    };
+                    await _employeeRepository.CreateUser(user);
+                    
+                }
             }
             catch (Exception ex)
-            {  
-                throw new InvalidOperationException("Error occurred while allocating leaves.", ex);
+            {
+                throw new InvalidOperationException("Error occurred while processing employee creation.", ex);
             }
-            return newEmployee;
+
+            return createdEmployee;
         }
+
 
         public async Task<string> DeleteEmployee(int id)
         {
+            if(id == null)
+            {
+                return "Requested ID not available ";
+            }
             var res = await _employeeRepository.DeleteEmployee(id);
             return res;
         }
